requirements.txt
aiopulsar-py==0.1.0a0
asgiref==3.7.2
attrs==22.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
coverage==6.5.0
faker==23.2.1
Flask==3.0.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
fastapi==0.89.1
greenlet==2.0.1
grpcio
grpcio-tools
idna==3.6
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.2.4
packaging==21.3
pluggy==1.0.0
protobuf==4.25.3
pulsar-client[avro]
py==1.11.0
PyDispatcher==2.0.6
pyparsing==3.0.9
pytest==7.1.3
pytest-flask
pytz==2024.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn[standard]==0.20.0
websockets==10.4
Werkzeug==3.0.1
